add tests for `cursor` method
